requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
transformers==4.35.2
scikit-learn==1.3.2
python-multipart==0.0.6
websockets==12.0
newsapi-python==0.2.7
pandas==2.1.3
numpy==1.26.2
pydantic==2.5.1
jinja2==3.1.2
beautifulsoup4==4.12.2
openai==1.3.5
httpx==0.25.2
